+++ smartdeckapp/lib/pages/register.dart
@@ -1,0 +1,127 @@
+import 'package:flutter/material.dart';
+
+class MyRegister extends StatefulWidget {
+  const MyRegister({Key? key}) : super(key: key);
+
+  @override
+  State<MyRegister> createState() => _MyRegisterState();
+}
+
+class _MyRegisterState extends State<MyRegister> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+          image: DecorationImage(image:  AssetImage('assets/background.jpg'),fit: BoxFit.cover)),
+      child: Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.transparent,
+          elevation: 0,
+        ),
+          body: Stack(
+            children: [
+              Container(
+                padding: EdgeInsets.only(left: 35, top: 130),
+                child: Text(
+                  'Create\nAccount',
+                  style: TextStyle(color: Colors.white, fontSize: 33),
+                ),
+              ),
+              SingleChildScrollView(
+                child: Container(
+                  padding: EdgeInsets.only(
+                      top: MediaQuery.of(context).size.height*0.5,
+                      right: 35,
+                      left: 35),
+                  child: Column(
+                    children: [
+                      TextField(
+                        decoration: InputDecoration(
+                            fillColor: Colors.grey.shade100,
+                            filled: true,
+                            hintText: 'First Name',
+                            border: OutlineInputBorder(
+                                borderRadius: BorderRadius.circular(10))),
+                      ),
+                      SizedBox(
+                        height: 30,
+                      ),
+                      TextField(
+                        decoration: InputDecoration(
+                            fillColor: Colors.grey.shade100,
+                            filled: true,
+                            hintText: 'Last Name',
+                            border: OutlineInputBorder(
+                                borderRadius: BorderRadius.circular(10))),
+                      ),
+                      SizedBox(
+                        height: 40,
+                      ),
+                      TextField(
+                        decoration: InputDecoration(
+                            fillColor: Colors.grey.shade100,
+                            filled: true,
+                            hintText: 'Email',
+                            border: OutlineInputBorder(
+                                borderRadius: BorderRadius.circular(10))),
+                      ),
+                      SizedBox(
+                        height: 40,
+                      ),
+                      TextField(
+                        obscureText: true,
+                        decoration: InputDecoration(
+                            hintText: 'Password',
+                            border: OutlineInputBorder(
+                                borderRadius: BorderRadius.circular(10))),
+                      ),
+                      SizedBox(
+                        height: 40,
+                      ),
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        children: [
+                          Text(
+                            'Sign Up',
+                            style: TextStyle(
+                                color: Colors.white,
+                                fontSize: 27,
+                                fontWeight: FontWeight.w700),
+                          ),
+                          CircleAvatar(
+                            radius: 30,
+                            backgroundColor: Colors.white,
+                            child: IconButton(
+                              color: Colors.white,
+                              onPressed: (){},
+                              icon: Icon(Icons.arrow_forward),
+                            ),
+                          )
+                        ],
+                      ),
+                      SizedBox(
+                        height: 40,
+                      ),
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        child:Button(
+                              onPressed: (){
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(builder: (context)=>MyLogin(),)
+                                )
+                              },
+                              )
+
+                        ],
+                      )
+                    ],
+                  ),
+                ),
+              )
+            ],
+          )
+      ),
+    );
+  }
+}
